refactor: rename dev to team
- semantic name change
- change to team as images folder
- all image references updated
- all references in team.js updatedd
