+++ src/data/2024/team.js
@@ -5,12 +5,12 @@ import Heba_Sayed from '@/data/2024/assets/images/dev/Heba_Sayed.jpg'
 import HudaHussaini from '@/assets/images/facilitators/Huda_Hussaini-x96.webp'
 import IrwinPatcharaput from '@/data/2024/assets/images/dev/Irwin_Patcharaput.jpeg'
 import JennaRitten from '@/assets/images/organizers/Jenna_Ritten-x96.webp'
-import JingranWang from '@/data/2025/assets/images/dev/Jingran_Wang-x96.webp'
+import JingranWang from '@/data/2025/assets/images/team/Jingran_Wang-x96.webp'
 import JuanLinaresOrihuela from '@/assets/images/facilitators/Juan_Linares_Orihuela-x96.webp'
 import MohamadAlabudi from '@/data/2024/assets/images/dev/Mohamad_Alabudi.jpg'
 import RahulVarmaAlluri from '@/data/2024/assets/images/dev/Rahul_Varma_Alluri.jpg'
-import SohanKetireddy from '@/data/2025/assets/images/dev/sohan_ketireddy-x96.webp'
-import SumiaSaleh from '@/data/2025/assets/images/dev/Sumia_Saleh-x96.webp'
+import SohanKetireddy from '@/data/2025/assets/images/team/sohan_ketireddy-x96.webp'
+import SumiaSaleh from '@/data/2025/assets/images/team/Sumia_Saleh-x96.webp'
 import TejasVaity from '@/data/2024/assets/images/dev/Tejas_Vaity.jpg'
 import ZavaarShah from '@/assets/images/facilitators/Zavaar_Shah-x96.webp'
