+++ src/data/2025/assets/images/team/Jingran_Wang-x96.webp
