+++ src/data/2025/assets/images/team/aaron_amano.webp
