+++ src/data/2025/assets/images/team/ali_elsharif.webp
