+++ src/data/2025/assets/images/team/dav_koziol.webp
