+++ src/data/2025/assets/images/team/greg_miller.png
