+++ src/data/2025/assets/images/team/justin_beaudry-x96.webp
