+++ src/data/2025/assets/images/team/justin_beaudry.png
