+++ src/data/2025/assets/images/team/maridy_mazaira.webp
