+++ src/data/2025/assets/images/team/rishiraj_sarkar.webp
