+++ src/data/2025/assets/images/team/shima_solati.webp
