+++ src/data/2025/assets/images/team/shugmi_shumunov-x96.webp
