+++ src/data/2025/assets/images/team/soham_mhatre.webp
