+++ src/data/2025/assets/images/team/sohan_ketireddy.webp
