+++ src/data/2025/team.js
@@ -1,24 +1,24 @@
-import AaronAmano from '@/data/2025/assets/images/dev/aaron_amano-x96.webp'
-import AliElSharif from '@/data/2025/assets/images/dev/ali_elsharif.webp'
-import DaveKoziol from '@/data/2025/assets/images/dev/dav_koziol.webp'
-import GregMiller from '@/data/2025/assets/images/dev/greg_miller-x96.webp'
-import HernandoMalabet from '@/data/2025/assets/images/dev/hernando_malabet.webp'
-import HodaSolati from '@/data/2025/assets/images/dev/hoda_solati.webp'
-import HudaHussaini from '@/data/2025/assets/images/dev/huda_hussaini.webp'
-import JennaRitten from '@/data/2025/assets/images/dev/jenna_ritten.webp'
-import JingranWang from '@/data/2025/assets/images/dev/Jingran_Wang-x96.webp'
-import JustinBeaudry from '@/data/2025/assets/images/dev/justin_beaudry-x96.webp'
-import KayleeBusenbark from '@/data/2025/assets/images/dev/kaylee_busenbark-x96.webp'
-import MaridyMazaira from '@/data/2025/assets/images/dev/maridy_mazaira.webp'
-import NisaSolano from '@/data/2025/assets/images/dev/nisa_solano.webp'
-import RaquelEstrada from '@/data/2025/assets/images/dev/raquel_estrada.webp'
-import Rehanashri from '@/data/2025/assets/images/dev/rehanashri-x96.webp'
-import RishirajSarkar from '@/data/2025/assets/images/dev/rishiraj_sarkar.webp'
-import ShimaSolati from '@/data/2025/assets/images/dev/shima_solati.webp'
-import ShugmiShumunov from '@/data/2025/assets/images/dev/shugmi_shumunov-x96.webp'
-import SohamMhatre from '@/data/2025/assets/images/dev/soham_mhatre.webp'
-import SohanKetireddy from '@/data/2025/assets/images/dev/sohan_ketireddy-x96.webp'
-import SumiaSaleh from '@/data/2025/assets/images/dev/Sumia_Saleh-x96.webp'
+import AaronAmano from '@/data/2025/assets/images/team/aaron_amano-x96.webp'
+import AliElSharif from '@/data/2025/assets/images/team/ali_elsharif.webp'
+import DaveKoziol from '@/data/2025/assets/images/team/dav_koziol.webp'
+import GregMiller from '@/data/2025/assets/images/team/greg_miller-x96.webp'
+import HernandoMalabet from '@/data/2025/assets/images/team/hernando_malabet.webp'
+import HodaSolati from '@/data/2025/assets/images/team/hoda_solati.webp'
+import HudaHussaini from '@/data/2025/assets/images/team/huda_hussaini.webp'
+import JennaRitten from '@/data/2025/assets/images/team/jenna_ritten.webp'
+import JingranWang from '@/data/2025/assets/images/team/Jingran_Wang-x96.webp'
+import JustinBeaudry from '@/data/2025/assets/images/team/justin_beaudry-x96.webp'
+import KayleeBusenbark from '@/data/2025/assets/images/team/kaylee_busenbark-x96.webp'
+import MaridyMazaira from '@/data/2025/assets/images/team/maridy_mazaira.webp'
+import NisaSolano from '@/data/2025/assets/images/team/nisa_solano.webp'
+import RaquelEstrada from '@/data/2025/assets/images/team/raquel_estrada.webp'
+import Rehanashri from '@/data/2025/assets/images/team/rehanashri-x96.webp'
+import RishirajSarkar from '@/data/2025/assets/images/team/rishiraj_sarkar.webp'
+import ShimaSolati from '@/data/2025/assets/images/team/shima_solati.webp'
+import ShugmiShumunov from '@/data/2025/assets/images/team/shugmi_shumunov-x96.webp'
+import SohamMhatre from '@/data/2025/assets/images/team/soham_mhatre.webp'
+import SohanKetireddy from '@/data/2025/assets/images/team/sohan_ketireddy-x96.webp'
+import SumiaSaleh from '@/data/2025/assets/images/team/Sumia_Saleh-x96.webp'
 
 export const teamData = [
   {
